requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.1
dnspython==2.4.2
gunicorn==21.2.0 